# gotham_like/docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Tessera Investigative Workbench",
  "description": "Local-first investigative intelligence & decision-support platform. Outputs are decision support: system inferences and analytical signals are never presented as established facts.",
  "version": "0.1.0"
 },
 "paths": {
  "/api/auth/login": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Login",
    "operationId": "login_api_auth_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LoginRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LoginResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/auth/logout": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Logout",
    "operationId": "logout_api_auth_logout_post",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Logout Api Auth Logout Post"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/auth/me": {
   "get": {
    "tags": [
     "auth"
    ],
    "summary": "Me",
    "operationId": "me_api_auth_me_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Me Api Auth Me Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/users": {
   "get": {
    "tags": [
     "users"
    ],
    "summary": "List Users",
    "operationId": "list_users_api_users_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "additionalProperties": true,
          "type": "object"
         },
         "type": "array",
         "title": "Response List Users Api Users Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "users"
    ],
    "summary": "Create User",
    "operationId": "create_user_api_users_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Create User Api Users Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/users/{user_id}": {
   "patch": {
    "tags": [
     "users"
    ],
    "summary": "Patch User",
    "operationId": "patch_user_api_users__user_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserPatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Patch User Api Users  User Id  Patch"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "LoginRequest": {
    "properties": {
     "username": {
      "type": "string",
      "maxLength": 64,
      "minLength": 1,
      "title": "Username"
     },
     "password": {
      "type": "string",
      "maxLength": 256,
      "minLength": 1,
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "username",
     "password"
    ],
    "title": "LoginRequest"
   },
   "LoginResponse": {
    "properties": {
     "access_token": {
      "type": "string",
      "title": "Access Token"
     },
     "token_type": {
      "type": "string",
      "title": "Token Type",
      "default": "bearer"
     },
     "expires_in": {
      "type": "integer",
      "title": "Expires In"
     },
     "user": {
      "additionalProperties": true,
      "type": "object",
      "title": "User"
     }
    },
    "type": "object",
    "required": [
     "access_token",
     "expires_in",
     "user"
    ],
    "title": "LoginResponse"
   },
...
   "Role": {
    "type": "string",
    "enum": [
     "ADMIN",
     "INVESTIGATOR",
     "ANALYST",
     "VIEWER"
    ],
    "title": "Role"
   },
...
   "UserCreate": {
    "properties": {
     "username": {
      "type": "string",
      "pattern": "^[a-z][a-z0-9_.-]{2,63}$",
      "title": "Username"
     },
     "display_name": {
      "type": "string",
      "maxLength": 200,
      "title": "Display Name",
      "default": ""
     },
     "password": {
      "type": "string",
      "maxLength": 256,
      "minLength": 12,
      "title": "Password"
     },
     "role": {
      "$ref": "#/components/schemas/Role"
     }
    },
    "type": "object",
    "required": [
     "username",
     "password",
     "role"
    ],
    "title": "UserCreate"
   },
   "UserPatch": {
    "properties": {
     "role": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Role"
       },
       {
        "type": "null"
       }
      ]
     },
     "is_active": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Is Active"
     },
     "display_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Display Name"
     }
    },
    "type": "object",
    "title": "UserPatch"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}